Added multi Threading
Defined the run() method called on start(): run->isup->vmstatus.
Defined up and running flags within the Host class to store the state of the node.
Threads are started and rejoined (only in the status command, not sure if that's ok. It's the first time I use the Threading library)
Adapted the rest of the code to the new Host flags
(minor) Rewritten the hostlist filter with 'node' option in a "pythoner" way, with list comprehension instead of two neasted for.

Now checking the status of all the nodes takes ~0.6s instad of ~16s.